feat: ask before configure and brew install steps

+++ setup.sh
@@ -1,40 +1,48 @@
 #!/bin/bash
 
 configure_macos() {
-    echo "Configuring macOS"
+    read -r -p "Do you want to enable dev friendly macOS configurations? (y/n) " answer
 
-    # "Finder: show all filename extensions"
-    defaults write NSGlobalDomain AppleShowAllExtensions -bool true
-    # "show hidden files by default"
-    defaults write com.apple.Finder AppleShowAllFiles -bool true
-    # "only use UTF-8 in Terminal.app"
-    defaults write com.apple.terminal StringEncodings -array 4
-    # "expand save dialog by default"
-    defaults write NSGlobalDomain NSNavPanelExpandedStateForSaveMode -bool true
-    # "show the ~/Library folder in Finder"
-    chflags nohidden ~/Library
-    # "Enable full keyboard access for all controls (e.g. enable Tab in modal dialogs)"
-    defaults write NSGlobalDomain AppleKeyboardUIMode -int 3
-    # "Enable subpixel font rendering on non-Apple LCDs"
-    defaults write NSGlobalDomain AppleFontSmoothing -int 2
-    # "Use current directory as default search scope in Finder"
-    defaults write com.apple.finder FXDefaultSearchScope -string "SCcf"
-    # "Show Path bar in Finder"
-    defaults write com.apple.finder ShowPathbar -bool true
-    # "Show Status bar in Finder"
-    defaults write com.apple.finder ShowStatusBar -bool true
-    # "Disable press-and-hold for keys in favor of key repeat"
-    defaults write NSGlobalDomain ApplePressAndHoldEnabled -bool false
-    # "Set a blazingly fast keyboard repeat rate"
-    defaults write NSGlobalDomain KeyRepeat -int 1
-    # "Set a shorter Delay until key repeat"
-    defaults write NSGlobalDomain InitialKeyRepeat -int 15
-    # "Enable tap to click (Trackpad)"
-    defaults write com.apple.driver.AppleBluetoothMultitouch.trackpad Clicking -bool true
-    # "Enable Safari’s debug menu"
-    defaults write com.apple.Safari IncludeInternalDebugMenu -bool true
-    # "Kill affected applications"
-    for app in Safari Finder Dock Mail SystemUIServer; do killall "$app" >/dev/null 2>&1; done
+    case ${answer:0:1} in
+    y | Y)
+        echo "Configuring macOS"
+        # "Finder: show all filename extensions"
+        defaults write NSGlobalDomain AppleShowAllExtensions -bool true
+        # "show hidden files by default"
+        defaults write com.apple.Finder AppleShowAllFiles -bool true
+        # "only use UTF-8 in Terminal.app"
+        defaults write com.apple.terminal StringEncodings -array 4
+        # "expand save dialog by default"
+        defaults write NSGlobalDomain NSNavPanelExpandedStateForSaveMode -bool true
+        # "show the ~/Library folder in Finder"
+        chflags nohidden ~/Library
+        # "Enable full keyboard access for all controls (e.g. enable Tab in modal dialogs)"
+        defaults write NSGlobalDomain AppleKeyboardUIMode -int 3
+        # "Enable subpixel font rendering on non-Apple LCDs"
+        defaults write NSGlobalDomain AppleFontSmoothing -int 2
+        # "Use current directory as default search scope in Finder"
+        defaults write com.apple.finder FXDefaultSearchScope -string "SCcf"
+        # "Show Path bar in Finder"
+        defaults write com.apple.finder ShowPathbar -bool true
+        # "Show Status bar in Finder"
+        defaults write com.apple.finder ShowStatusBar -bool true
+        # "Disable press-and-hold for keys in favor of key repeat"
+        defaults write NSGlobalDomain ApplePressAndHoldEnabled -bool false
+        # "Set a blazingly fast keyboard repeat rate"
+        defaults write NSGlobalDomain KeyRepeat -int 1
+        # "Set a shorter Delay until key repeat"
+        defaults write NSGlobalDomain InitialKeyRepeat -int 15
+        # "Enable tap to click (Trackpad)"
+        defaults write com.apple.driver.AppleBluetoothMultitouch.trackpad Clicking -bool true
+        # "Enable Safari’s debug menu"
+        defaults write com.apple.Safari IncludeInternalDebugMenu -bool true
+        # "Kill affected applications"
+        for app in Safari Finder Dock Mail SystemUIServer; do killall "$app" >/dev/null 2>&1; done
+        ;;
+    *)
+        echo "Skipping Configuration."
+        ;;
+    esac
 }
 
 install_brew() {
@@ -52,6 +60,20 @@ install_packages() {
     brew bundle check --global --verbose
 }
 
+install_brew_and_packages() {
+    read -r -p "Do you want to install Homebrew and recommended packages? (y/n) " answer
+
+    case ${answer:0:1} in
+    y | Y)
+        install_brew
+        install_packages
+        ;;
+    *)
+        echo "Skipping Brew Packages."
+        ;;
+    esac
+}
+
 setup_symlinks() {
     echo "Setting up symlinks"
     ln -sfn "$PWD/config/aerospace" "$HOME/.config/"
@@ -60,8 +82,7 @@ setup_symlinks() {
 
 setup_mac2k() {
     configure_macos
-    install_brew
-    install_packages
+    install_brew_and_packages
     setup_symlinks
     cd "$PWD/dots2k" && ./setup.sh && cd .. || exit 1
 }
